Changes to get compilation with debugging symbols defines (--enable-debug argument to configure): - Support --enable-debug=full (treated as --enable-debug) - Define _PLAYTEST as well as _DEBUG - Add includes - Guarded Win32 code - Clarified arguments to DPRINTF - Sorted initializations - Fixed casts - Fixed case of includes - Changed NULL to 0 when used in arithmetic - Commented additional tokens after #endif - Added SDL alternatives to Win32 code - Added return statement for non-void function SDLMessageHandler - Replaced multi-character constants - Added signing casts
git-svn-id: http://ctp2.darkdust.net/anonsvn/branches/linux@434 34d2fc0c-33df-0310-8a34-85f8040ba873

+++ ctp2_code/ai/mapanalysis/mapgrid.h
@@ -131,7 +131,7 @@ public:
 			     (xy_pos.x / m_resolution);
 
 		Assert(elem >= 0);
-		Assert(elem < m_values.size());
+		Assert((unsigned) elem < m_values.size());
 		
 		
 		Assert(((double) m_values[elem] + value) == (double) (m_values[elem] + value));
@@ -404,7 +404,7 @@ public:
 		elem = ((xy_pos.y / m_resolution) * m_xGridSize) + 
 			    (xy_pos.x / m_resolution);
 		Assert(elem >= 0);
-		Assert(elem < m_values.size());
+		Assert((unsigned) elem < m_values.size());
 
 		m_values[elem] = value;
 	}
@@ -447,7 +447,7 @@ public:
 	{
 		std::ostringstream ost;
 		
-		for (sint32 elem = 0; elem < m_values.size(); elem++)
+		for (sint32 elem = 0; (unsigned) elem < m_values.size(); elem++)
 		{	
 				ost.width( 3);
 				ost << m_values[elem];
